init
First commit now that it is working. This sets some params and looks up the user. All the games are saved in a json games array. Each one can be accessed. This version just prints out the number of games and errors. Sometimes the JSON is corrupted. 20 games is the max.